Extract default effects and update default decks
