+++ src/data/example-decks/deck-20.ts
@@ -7,18 +7,18 @@ export const exampleDeck20: ExampleDeckTemplate = {
     name: "Beispieldeck 20 Spieler",
     pickedRoles: {
         werwolf: 4,
-        dorfbewohner: 4,
+        dorfbewohner: 3,
         seherin: 1,
         hexe: 1,
         jaeger: 1,
         heiler: 1,
         prinz: 1,
-        cupido: 1,
+        armor: 1,
         priest: 1,
         wolfsjunges: 1,
-        wolfsseher: 1,
-        whiteWolf: 1,
-        vampir: 1,
+        traumwolf: 1,
+        putzfrau: 1,
+        vampir: 2,
         einsamerwerwolf: 1,
     },
 }
